svg/text/parsing/shape-margin-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Text: parsing shape-margin with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/text.html#ShapeMarginProperty" rel="help"/><h:meta content="shape-margin supports the full grammar 'length-percentage'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("shape-margin", "0", "0px");
test_valid_value("shape-margin", "10px");
test_valid_value("shape-margin", "20%");
test_valid_value("shape-margin", "calc(2em + 3ex)");

  ]]></script></svg>
